html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-pointer-move-pointer-after-out.html

<!doctype html>
<title>pointer updates (moveBefore pointerAfter out of the media element)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<div id="holder"></div>
<script>
var events = [];
var a = 0;
var b = 0;
var c = 0;
var d = 0;
</script>
<video id="v"><source id="source-a" onerror="a++; events.push('a')"><source id="source-b" onerror="b++; events.push('b')" src="resources/delayed-broken-video.py"><source id="source-c" onerror="c++; events.push('c')"><source id="source-d" onerror="d++; events.push('d')"></video>
<script>
async_test(function(t) {
  var video = document.getElementById('v');
  assert_implements(typeof video.moveBefore == 'function', 'ParentNode.moveBefore() is required for this test');

  // Move pointerAfter out while b is pending. d should become the next candidate.
  document.getElementById('holder').moveBefore(document.getElementById('source-c'), null);

  window.onload = t.step_func(function() {
    assert_equals(a, 1, 'error events on a');
    assert_equals(b, 1, 'error events on b');
    assert_equals(c, 0, 'error events on c');
    assert_equals(d, 1, 'error events on d');
    assert_array_equals(events, ['a', 'b', 'd'], 'processed source order');
    t.done();
  });
});
</script>
